requirements.txt
aspy.yaml==1.3.0
certifi==2019.9.11
cfgv==2.0.1
chardet==3.0.4
coverage==4.5.4
coveralls==1.8.2
Django==2.2.6
django-rest-framework==0.1.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docopt==0.6.2
identify==1.4.7
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
nodeenv==1.3.3
pre-commit==1.20.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
urllib3==1.25.6
virtualenv==16.7.7
zipp==0.6.0
